qcacld-3.0: Change data type of the code for mac trace to uint16_t
Curretly data type of code for mac trace is uint8_t and
qdf trace expects the data type of code as uint16_t because of
which it may result in incorrect code type for the messages with
message id in uint16_t.

To address above issue change the type of the mac trace to
uint16_t.

CRs-Fixed: 2719765

+++ core/mac/inc/mac_trace.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013-2016, 2018-2019 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2013-2016, 2018-2020 The Linux Foundation. All rights reserved.
  *
  * Permission to use, copy, modify, and/or distribute this software for
  * any purpose with or without fee is hereby granted, provided that the
@@ -44,7 +44,7 @@
  *
  * Return: None
  */
-static inline void mac_trace(struct mac_context *mac_ctx, uint8_t code,
+static inline void mac_trace(struct mac_context *mac_ctx, uint16_t code,
 			     uint16_t session, uint32_t data)
 {
 	qdf_trace(QDF_MODULE_ID_PE, code, session, data);
